Reorganize some of the code.
Put the kernel/LocalMem etc. into a new low_level module so it is clear
that this is low-level functionality that is not cross-backend capable.
